Added Direction Enums to Scarf Hook
This simplifies usage of scarf in the future.
In addition made hitbox of box scale a little with length to lessen risk of missing.
When throwing scarf upwards it gains a bit of width.
When throwing scarf sideways it gains a bit of height.

So both become more 'cone' like.